+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v0.11.1
 
+- Add `deselectOnDblClick` and `deselectOnEscape` to allow disabling deselection on double click or escape when set to `false` (#31)
 - Fix an issue updating the x and y scale domains (#30)
 
 ## v0.11.0